Update ImageRegression_attempt1.ipynb
Updated code on the training/testing data set; only ran the model for 4 classes (age 0 to 3). Added a variable called num_class so that we can easily change how many agess we want to classify. I only ran 10 epochs but more can be run for the actual model. Still need to print a confusion matrix and/or heat map and also need to consider pre-processing the data using data augmentation to increases the accuracy so that it is higher than 0.47.